Ref: main page refactored #129
