+++ src/components/main/CardCarousel.jsx
